fix: check state and remove auth middleware
